Paperspace test!

+++ src/data.py
@@ -184,6 +184,6 @@ def test_chips(imgs, mask,
                 mask_chips += [(temp_mask > 0).astype(float)[..., np.newaxis]]
 
     img_chips = np.array(img_chips).astype(float)
-    mask_chips = np.array(mask_chips).astype(float)
+    mask_chips = np.array(mask_chips)
 
     return img_chips, mask_chips
